+++ jest.config.specific.js
@@ -5,12 +5,6 @@ module.exports = {
         '^@/(.*)$': '<rootDir>/src/$1',
     },
     testEnvironment: 'node',
+    testMatch: ['**/__tests__/integration.test.(js|ts)?(x)'],
     moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-    coverageReporters: ['lcov', 'text-summary'],
-    coverageDirectory: 'coverage',
-    coverageThreshold: {
-        global: {
-            lines: 100,
-        },
-    },
 };
