+++ src/shared/database/__tests__/index.test.ts
@@ -2,14 +2,6 @@
 // @ts-nocheck
 import { Database } from '..';
 
-const mockInitializeApp = jest.fn();
-const mockCert = jest.fn();
-
-jest.mock('firebase-admin/app', () => ({
-    initializeApp: () => mockInitializeApp(),
-    cert: () => mockCert(),
-}));
-
 jest.mock('firebase-admin/firestore', () => ({
     getFirestore: jest.fn(() => ({
         collection: jest.fn(() => ({
